abstracted signing operations into their own class
