+++ functions/signing/.gitignore
@@ -1,10 +1,10 @@
 *
 !/main.py
-!/requirements.txt
 !/.gitignore
 !/.apexignore
 !/README.md
-!/signing.py
+!/signer.py
+!/signing_handler.py
 !/aws_decrypt.py
 !/__init__.py
 !/uW-sk.enc
